Add more complex rule result evaluations
This adds a small parser to allow us to compare more
complex rules when evaluating results.

e.g. we could now say something like:

```
default_result: PASS or INCONSISTENT
```

Which could be used for rules that are known to fail due to known issues
in OpenShift.

Right now the syntax allows for the following:

```
FOO

FOO or BAR

FOO or BAR or BAZ
```

This doesn't currenlt validate that we're setting correct and specific
results as defined by Compliance Operator. Though this could be easily
added to the parser.

Is this over-engineered? Yes.
Was a parser like this needed? Probably not

But it was sure fun to code :)

The main use-case is a rule that's been failing every once in a while
with INCONSISTENT... It's already reported [1], but it gets no attention
nor do I expect a fix soon.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2001442

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>