helpers: Enable showNotApplicable checks
Several versions ago, CO started defaulting to not creating check
results for rules that are not applicable. While that's good for the end
user, doing so in the e2e test environment means that we completely miss
checks that have become not applicable for a platform for one reason or
another. We should always evaluate all test results against the expected
results.

Jira: CMP-1417