helpers: Fail if there is no test result YAML file for current version, but exists for other versions
Right now, the CaC e2e tests do support versioned test results, but if a
test result does not exist for a specific version, the tests simply fall
back to the version-less test result and if that doesn't exist either,
the result is silently ignored.

This means that once we switch to a new version of OCP but forgot to add
a versioned test result for that version, the check is silently ignored.

We need to have a mechanism that would alert us when we switch to a new
version but don't add a versioned result for that version.

Jira: CMP-1415