Add basic scan summaries to test logs
Since we have the data available to use when we run the tests, we can
log a basic summary about how many checks were in specific states. This
can be useful when you want to get a quick idea of where a profile is
with respect to its pass/fail rate.