Generalize e2e scan settings
Previously, the e2e suite would create a scan setting that always
auto-remediated any findings. While this is still useful, we don't want
to automatically remediate when we're testing on managed platforms since
we're not supporting that initially.

This commit generalizes the e2e scan settings so that they don't
automatically remediate if the suite is configured to bypass
remediations (which is what we do when testing managed offerings).