Add more details to remediation timeout
Right now, when a remediation times out, the test suite says the command
failed, but only includes a reference to the script that contains the
remediation.

This commit adds the command output and the error so that it's easier to
understand what's going on with the script and why it might be broken.