Increase timeout for applying manual remediations
We're seeing timeouts now that we're using oc adm
wait-for-stable-cluster as part of the logic to ensure manual
remediations are applied to the cluster.

This commit bumps the context timeout to 1 hour to give the cluster more
time to bounce the cluster operators after the manual remediations are
applied. Otherwise we end up with errors similar to the following:

  Error: context deadline exceeded