CMP-3385: Refactor test suite to group rules
With the amount of profiles we're testing across OpenShift versions, we
experience a lot of duplicate failures if one rule breaks because it
will affect multiple profiles across multiple OpenShift versions.

To make testing easier to manage, we want to try building a single
tailored profile for all platform rules and another for node rules. Then
run all rules of the same type together. After that, we'll check the
outcome of each rule like we did.

The result should be that we only need to run a single test to exercise
all platform rules on a single OpenShift version. Instead of running
multiple platform profiles on each OpenShift version.