Remove assert profile function
We're changing the test strategy so that we don't need to test each
profile individually. Instead, we want to stitch together rules into a
single profile and run them all at once so we reduce the amount of churn
when a single rule fails across multiple profiles.

This commit removes the assertProfile assertion so that it's not
required.