Add additional test utilities to helper package
These functions are similar to what we already have in helpers.go, but
uses a different interface so that we can continue using the old ones
while we move things over. Eventually, we can remove the versions of
these functions in helpers.go and just rely on the ones in the helpers
package.