Refactor rule lookups
We need to make sure the tailored profile only contains rules from the
content we're testing, which is coming from a container image built
against the latest master in CaC/content. Otherwise, if we rely on the
default bundle it could fail because the rule doesn't exist there
anymore.

This means we need to make sure we filter the rule lookup by the
profile bundle name we're using in testing (e2e-ocp4, e2e-rhcos4).