Separate node rules based on product
We can't put all node rules in a single profile because they span
multiple products, and the TailoredProfile validation prohibits that
(probably because we need all the rules in a single datastream to invoke
the scan, and datastreams are built on a per-product basis).

This commit separates the node profile into one for RHCOS rules and one
for OpenShift node rules before creating the profiles and invoking the
scans.