Add remediation workflow
Make it so the suite applies remediations so we can test the ones we
ship in the Compliance Operator. Also add some additional logging for
remediations that aren't working as expected so we have information we
use can to investigate.

Also rescan the environment after the remediations are applied.

Subsequent PRs will add support to check the latest scan results to
expected values after the remediations are applied.