Use a reasonable default for -content-image
We had this defined as an env in the Makefile, and passed it into the
suite as a test argument, then checked if it was always set.

We can simplify this by just setting a reasonable default. We can still
let people provide their own content images, but by just having a
default value we can remove unnecessary variables from the Makefile and
simplify the validation logic in the test suite.