Add wait logic for rescanning
If we rescan the environment and then immediately start waiting for the
scan to complete, there is a race condition where the wait logic will
see that the scan is already done since the controller logic hasn't
updated the status, yet.

This commit updates that so that we check if we're rescanning the
environment or not, and waits for the suite to be updated.