Safely wait for scan suites to get created
We had a race condition where the tests would fail because of how
quickly the test would check for the compliance suite after the binding
was created. The binding controller needs time to actually create the
binding, which creates the compliance suite.

This commit makes the WaitForComplianceSuite function smarter about a
compliance suite that wasn't found, and accepting that it's probably
just the first scan and that it should poll anyway. Doing this as
opposed to a sleep().