Rescan until remediation dependencies are met
Since remediations can have a dependency relationship with other
remediations, we need to make sure we apply them, and rescan so that
dependencies that are waiting for the initial remediations to be applied
can have their turn.

Abstract this into a separate utility function and make sure we run one
final scan after it's all done resolving remediation dependencies.