Add scan cleanup logic
If you're using the suite against a cluster and you rerun the tests, a
leftover scan binding from the previous run will throw off the test.

Instead, let's detect if the scan binding exists, delete it, and wait
for the resources associated with that scan to get deleted before we
create a new scan setting binding for fresh results.