Tighten e2e targets for platform and node
Running `make e2e-platform` and `make e2e-node` would also invoke the
profile tests, which isn't necessarily what we want since we don't
guarantee the profile will be passed in. Instead, we can tighten the
testing expression so that it only runs the tests we care about for
those cases.