Update ValidateProfile to use full profile name with product
Profiles are namespaced by the product for uniqueness, since there can
be a `moderate` profile in the ocp4 profile bundle and the rhcos4
profile bundle.

This commit updates the profile testing so that it concatenates the product
and profile for the actual profile name before the lookup. Otherwise,
the test just fails because the `moderate` profile doesn't exist (it's
either `ocp4-moderate` or `rhcos4-moderate`).