CMP-3385: Reduce time to remediate during tests
Increase MaxUnavailable for worker and master MachineConfigPools so that
we don't have to wait so long for the remediations to apply. Some
remediations have dependencies, where one need to be applied before
another, and that results in multiple cluster reboots.

This commit increases the MaxUnavailable policy to 100%, meaning we're
accepting the fact that the master nodes might be down for a while, in
favor of speeding up the tests.