Add parameter for logging directory
We want to improve the overall experience for contributors debugging CI
results. To do that, we can generate assertion files and mismatch files
that show exactly what failed. But, we also want to make sure we can
still run the suite locally, and not assume /logs/artifacts/ is always
available.

This commit adds a test parameter so that users can pass in the log
directory of their choosing. By default, the suite will write logs to
/logs/artifacts, which we can easily fish out of prow.