CMP-3385: Improve test logging and result output
This commit refactors how results are checked against assertion files.
It pulls the assertion file generation up to the e2e suite, instead of
keeping it buried in the utilities. It also saves the results and
mismatches from each run in their own files that we can dig them out of prow
to make debugging failures easier. This is better than sifting through
plain text logs because the data is structured and isolated to specific
cases (e.g., a file for the initial results, and a file for the final
results).

This also included changing the e2e suite so that it stashed the results
after a scan, instead of fetching the ComplianceCheckResults in the
utilities. This made some of the comparison easier since we didn't need
to deal with a CRD, just simpler data types like strings for
evaluation.

Also include trivial linting fixes that slipped in since we don't have
linting enabled on each PR (something we need to do to maintain
cleanliness).