CMP-3385: Reimplement manual remediations
The refactor to reorganize the e2e suite eliminated the ability to apply
manual remediations. This commit adds that back into the suite at an
appropriate place. It also introduces a new test configuration option
that let's callers control how long they want to wait for the
remediation timeout. It defaults to 30 minutes per remediation, which
run in parallel.