Share the test configuration across tests
The TestMain function created the initial test configuration, and does
important things like cloning the content repository and installing the
operator. However, each test case was creating it's own test
configuration. This is mostly fine, but doesn't safely account for the
ContentDir argument since that's updated in Setup(), which is only ever
called from TestMain(). All the other test functions assume Setup() will
handle the content directory, but that doesn't work if they don't share
the same test configuration.

This commit moves the test configuration to a global that setup creates
and bootstraps using the parse test arguments. From there, each of the
test cases can use the global test configuration, so long as it's not
modified (it should be a read-only resource after Setup() so that it
doesn't cause inconsistent behaviors between tests).