Pull assertion files out of the utilities
Previously, we had a utility that would assert the scan results against
a set of assertions, and if the assertions didn't exist it would
generate an assertion file. We can pull this logic up to the test layer,
where it defines the assertion file it needs for the test, and it's also
responsible for writing the assertions for failed tests to disk. Then,
the verification utility methods are only responsible for matching
expected and actual results.