Simplify profile bundle setup
Previously, the test suite would always load in the default profile
bundle available with the compliance operator, and it would also load in
the profile bundles from an optional argument for ContentImage. This
ContentImage gives us the flexibility to pass in content images from PR
builds, or content images from local development. The ContentImage also
has a reasonable default value that points to the latest content image
built upstream in ComplianceAsCode/content, which is what the operator
uses by default when deployed from the master/main branch.

This commit simplifies the setup by checking if the ContentImage is
overridden. If it is, then it deletes the standard profile bundles and
replaces them with the supplied content image from CI, or local
development, with the default profile bundle names (ocp4, rhcos4). If
the user hasn't passed in a different ContentImage, then we just use
what's available by default and don't bother loading the test config
ContentImage since it's already being installed by the operator anyway,
saving some setup time.