Fix controller-runtime logging setup
The test suite uses the k8s client packages for fetching resources,
which use logging utilities from controller-runtime. The suite should
setup the logger for the client. To do that we need to include the
dependency, and use it appropriately in the tests, otherwise we'll get
issues like the following:

  [controller-runtime] log.SetLogger(...) was never called; logs will not be displayed.