Use t.Error instead of t.Fatal for assertion checking
The tests will fail if there are any discrepancies between the expected
cluster state and the observed state. Using t.Fataf() or t.Fatal marks
the test as failed, which is what we want, but it also stop execution of
the test. Since we have some helper methods that process the mismatched
results to generate assertion files and reports, we don't want to
completely stop execution so that we can still produce those files.

This commit updates those specific parts of the tests to use Error
instead of Fatal so we perserve the behavior to mark the test as failed,
but continue processing in the test to cleanup and generate reports.